+++ inc/ti/fn/fnisenum.h
@@ -16,9 +16,3 @@ static int do__f_is_enum(ti_query_t * query, cleri_node_t * nd, ex_t * e)
 
     return e->nr;
 }
-
-static int do__f_isenum(ti_query_t * query, cleri_node_t * nd, ex_t * e)
-{
-    log_error("function `isenum` is deprecated, use `is_enum` instead");
-    return do__f_is_enum(query, nd, e);
-}