+++ inc/ti/fn/fniserr.h
@@ -16,9 +16,3 @@ static int do__f_is_err(ti_query_t * query, cleri_node_t * nd, ex_t * e)
 
     return e->nr;
 }
-
-static int do__f_iserr(ti_query_t * query, cleri_node_t * nd, ex_t * e)
-{
-    log_error("function `iserr` is deprecated, use `is_err` instead");
-    return do__f_is_err(query, nd, e);
-}