+++ inc/ti/fn/fnisinf.h
@@ -31,9 +31,3 @@ static int do__f_is_inf(ti_query_t * query, cleri_node_t * nd, ex_t * e)
 
     return e->nr;
 }
-
-static int do__f_isinf(ti_query_t * query, cleri_node_t * nd, ex_t * e)
-{
-    log_error("function `isinf` is deprecated, use `is_inf` instead");
-    return do__f_is_inf(query, nd, e);
-}