+++ inc/ti/fn/fnisnan.h
@@ -27,9 +27,3 @@ static int do__f_is_nan(ti_query_t * query, cleri_node_t * nd, ex_t * e)
 
     return e->nr;
 }
-
-static int do__f_isnan(ti_query_t * query, cleri_node_t * nd, ex_t * e)
-{
-    log_error("function `isnan` is deprecated, use `is_nan` instead");
-    return do__f_is_nan(query, nd, e);
-}