+++ inc/ti/fn/fnisraw.h
@@ -16,9 +16,3 @@ static int do__f_is_raw(ti_query_t * query, cleri_node_t * nd, ex_t * e)
 
     return e->nr;
 }
-
-static int do__f_israw(ti_query_t * query, cleri_node_t * nd, ex_t * e)
-{
-    log_error("function `israw` is deprecated, use `is_raw` instead");
-    return do__f_is_raw(query, nd, e);
-}