+++ inc/ti/fn/fnisstr.h
@@ -16,9 +16,3 @@ static int do__f_is_str(ti_query_t * query, cleri_node_t * nd, ex_t * e)
 
     return e->nr;
 }
-
-static int do__f_isstr(ti_query_t * query, cleri_node_t * nd, ex_t * e)
-{
-    log_error("function `isstr` is deprecated, use `is_str` instead");
-    return do__f_is_str(query, nd, e);
-}