+++ inc/ti/fn/fnisthing.h
@@ -16,9 +16,3 @@ static int do__f_is_thing(ti_query_t * query, cleri_node_t * nd, ex_t * e)
 
     return e->nr;
 }
-
-static int do__f_isthing(ti_query_t * query, cleri_node_t * nd, ex_t * e)
-{
-    log_error("function `isthing` is deprecated, use `is_thing` instead");
-    return do__f_is_thing(query, nd, e);
-}