+++ inc/ti/fn/fnistuple.h
@@ -16,9 +16,3 @@ static int do__f_is_tuple(ti_query_t * query, cleri_node_t * nd, ex_t * e)
 
     return e->nr;
 }
-
-static int do__f_istuple(ti_query_t * query, cleri_node_t * nd, ex_t * e)
-{
-    log_error("function `istuple` is deprecated, use `is_tuple` instead");
-    return do__f_is_tuple(query, nd, e);
-}